+++ react-exercises/simple-anime-react/src/components/simple-anime-lib/effects/fade-in.js
@@ -1,5 +1,5 @@
 import React, { useRef, memo } from 'react';
-import { DEFAULT_FADEOUT_CONFIG } from './constants';
+import { DEFAULT_FADEIN_CONFIG } from '../constants';
 import {
   useAnimationFrame,
   timingFunction,
@@ -7,27 +7,26 @@ import {
   getDuration,
   getIterationCount,
   getEasing,
-} from './helper';
+} from '../helper';
 
-const FadeOut = memo((props) => {
-  const delay = getDelay(props.delay, DEFAULT_FADEOUT_CONFIG.delay);
-  const duration = getDuration(props.duration, DEFAULT_FADEOUT_CONFIG.duration);
+const FadeIn = memo((props) => {
+  const delay = getDelay(props.delay, DEFAULT_FADEIN_CONFIG.delay);
+  const duration = getDuration(props.duration, DEFAULT_FADEIN_CONFIG.duration);
   const iterationCount = getIterationCount(
     props.iterationCount,
-    DEFAULT_FADEOUT_CONFIG.iterationCount
+    DEFAULT_FADEIN_CONFIG.iterationCount
   );
-  const easing = getEasing(props.easing, DEFAULT_FADEOUT_CONFIG.easing);
+  const easing = getEasing(props.easing, DEFAULT_FADEIN_CONFIG.easing);
   const timing = timingFunction[easing];
-  const elementRef = useRef({ style: { opacity: 1.0 } });
+  const elementRef = useRef({ style: { opacity: 0.0 } });
 
   const draw = (progress) => {
     if (elementRef.current !== null) {
-      elementRef.current.style.opacity = 1 - progress;
+      elementRef.current.style.opacity = progress;
     }
   };
 
   useAnimationFrame({ delay, duration, timing, draw, iterationCount });
-
   return (
     <span ref={elementRef} style={elementRef.current.style}>
       {props.children}
@@ -35,4 +34,4 @@ const FadeOut = memo((props) => {
   );
 });
 
-export { FadeOut };
+export { FadeIn };
