+++ react-exercises/simple-anime-react/src/components/simple-anime-lib/effects/fade-out.js
@@ -1,5 +1,5 @@
 import React, { useRef, memo } from 'react';
-import { DEFAULT_FADEIN_CONFIG } from './constants';
+import { DEFAULT_FADEOUT_CONFIG } from '../constants';
 import {
   useAnimationFrame,
   timingFunction,
@@ -7,26 +7,27 @@ import {
   getDuration,
   getIterationCount,
   getEasing,
-} from './helper';
+} from '../helper';
 
-const FadeIn = memo((props) => {
-  const delay = getDelay(props.delay, DEFAULT_FADEIN_CONFIG.delay);
-  const duration = getDuration(props.duration, DEFAULT_FADEIN_CONFIG.duration);
+const FadeOut = memo((props) => {
+  const delay = getDelay(props.delay, DEFAULT_FADEOUT_CONFIG.delay);
+  const duration = getDuration(props.duration, DEFAULT_FADEOUT_CONFIG.duration);
   const iterationCount = getIterationCount(
     props.iterationCount,
-    DEFAULT_FADEIN_CONFIG.iterationCount
+    DEFAULT_FADEOUT_CONFIG.iterationCount
   );
-  const easing = getEasing(props.easing, DEFAULT_FADEIN_CONFIG.easing);
+  const easing = getEasing(props.easing, DEFAULT_FADEOUT_CONFIG.easing);
   const timing = timingFunction[easing];
-  const elementRef = useRef({ style: { opacity: 0.0 } });
+  const elementRef = useRef({ style: { opacity: 1.0 } });
 
   const draw = (progress) => {
     if (elementRef.current !== null) {
-      elementRef.current.style.opacity = progress;
+      elementRef.current.style.opacity = 1 - progress;
     }
   };
 
   useAnimationFrame({ delay, duration, timing, draw, iterationCount });
+
   return (
     <span ref={elementRef} style={elementRef.current.style}>
       {props.children}
@@ -34,4 +35,4 @@ const FadeIn = memo((props) => {
   );
 });
 
-export { FadeIn };
+export { FadeOut };
